make webgate connect users to a random IRC server in a subset
the need for this appeared due to the fact we weren't allowed to run an authorized webgate and therefore had to connect to IRC as a normal user. not surprisingly this introduced an issue with hitting the simultaneous connections limit from a single IP (ours), and the obvious solution can be simply to try sending webgate users to different IRC servers in the IRC network, while trying to implement a naive balancing them between the servers. this design is indeed very ugly, but it has proven to work fine for several years already