check users connection and prematurely QUIT them from IRC when they've gone
the reason for this hack is our IRC network has HUGE ping timeouts set
(like 10 minutes) and as currently we can't know when the user has left if
he closes the browser, his dead "shadow" is pretending to be "online" for
another whopping 10 minutes until IRC's "ping timeout" finally kicks in.

and this is awful.

so this changeset adds a scheduled pings into the websocket connection
(4 times a minute), and if we receive no pings from a user in 75 seconds,
we can assume he's already left, and then we can simply disconnect him
from the chat ourselves - and it would be nearly 10 times quicker than
the IRC disconnecting him by an own timeout.

also here we'll try to make user unobtrusively ping his own nick on IRC
(once in 3 minutes and only if the user has sent nothing during this period).
i've found that thanks to this i'm getting much less random disconnects
from the server than before. but maybe that's because the IRC i'm using
is badly broken - it should PING/PONG us once in a while itself, i know.